+++ CHANGELOG.rst
@@ -15,6 +15,13 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.6
+---
+2019-10-03 (not yet released)
+
+- Simple search query backend.
+- Deprecating old search backend.
+
 0.5
 ---
 2019-09-29